feat(phase15e): gem consumer independence (worker.entrypoint, build CLI, scaffolder)
- Split runtime into worker_module.mjs; thin worker.mjs; track build/worker.entrypoint.mjs for wrangler main
- Add cloudflare-workers-build (homurabi + --standalone); git mv compile-erb, compile-assets, patch-opal-evals into gem
- Add cloudflare-workers-new + project templates; standalone build pulls homurabi vendor/ via Gemfile path for Opal
- Pin opal = 1.8.3.rc1 in gemspecs; add docs/ARCHITECTURE.md; update examples and root README

Made-with: Cursor
