+++ Rakefile
@@ -1,17 +1,21 @@
 # frozen_string_literal: true
 
 # Phase 15-A — thin wrappers around the generic build CLIs.
-# `npm run build` remains the primary entry point; these tasks exist
-# for gem / CI documentation parity (see docs/TOOLCHAIN_CONTRACT.md).
+# `npm run build` / `bundle exec cloudflare-workers-build` is the primary entry point.
+
+GEM_ERB = 'gems/cloudflare-workers-runtime/exe/compile-erb'
+PATCH   = 'gems/cloudflare-workers-runtime/runtime/patch-opal-evals.mjs'
 
 desc 'Compile ERB templates into build/homurabi_templates.rb'
 task 'build:erb' do
-  sh 'ruby', 'bin/compile-erb', '--input', 'views', '--output', 'build/homurabi_templates.rb', '--namespace', 'HomurabiTemplates'
+  sh 'ruby', GEM_ERB, '--input', 'views', '--output', 'build/homurabi_templates.rb', '--namespace', 'HomurabiTemplates'
 end
+
+GEM_ASSETS = 'gems/cloudflare-workers-runtime/exe/compile-assets'
 
 desc 'Embed public/ assets into build/homurabi_assets.rb'
 task 'build:assets' do
-  sh 'ruby', 'bin/compile-assets', '--input', 'public', '--output', 'build/homurabi_assets.rb', '--namespace', 'HomurabiAssets'
+  sh 'ruby', GEM_ASSETS, '--input', 'public', '--output', 'build/homurabi_assets.rb', '--namespace', 'HomurabiAssets'
 end
 
 desc 'Compile Opal bundle (same argv/env contract as npm run build:opal)'
@@ -19,7 +23,6 @@ task 'build:opal' do
   out = ENV.fetch('HOMURABI_OPAL_OUTPUT', 'build/hello.no-exit.mjs')
   inp = ENV.fetch('HOMURABI_OPAL_INPUT', 'app/hello.rb')
   patch = ENV.fetch('HOMURABI_OPAL_PATCH_INPUT', out)
-  # argv form avoids shell interpolation on paths that may contain spaces.
   File.open('build/opal.stderr.log', 'w') do |err_io|
     ok = system(
       { 'OPAL_PREFORK_DISABLE' => '1' },
@@ -32,10 +35,12 @@ task 'build:opal' do
     )
     fail 'opal compile failed (see build/opal.stderr.log)' unless ok
   end
-  sh 'node', 'bin/patch-opal-evals.mjs', patch
+  sh 'node', PATCH, patch
 end
 
-desc 'ERB + assets + Opal (matches npm run build)'
-task build: %w[build:erb build:assets build:opal]
+desc 'Full homurabi build (ERB + assets + Opal + patch + worker.entrypoint.mjs)'
+task :build do
+  sh 'bundle', 'exec', 'cloudflare-workers-build'
+end
 
 task default: :build
