+++ examples/classic-top-sinatra/Gemfile
@@ -2,6 +2,7 @@
 source 'https://rubygems.org'
 ruby '>= 3.4.0'
 
+# Mode A — monorepo dev | Mode B — `gem install --local` + path to worktree (see `examples/minimal-sinatra/Gemfile`).
 gem 'opal', path: '../../vendor/opal-gem'
 gem 'cloudflare-workers-runtime', path: '../../gems/cloudflare-workers-runtime'
 gem 'sinatra-cloudflare-workers', path: '../../gems/sinatra-cloudflare-workers'
