+++ examples/classic-top-sinatra/wrangler.toml
@@ -1,4 +1,4 @@
 name = "classic-top-sinatra-example"
-main = "../../gems/cloudflare-workers-runtime/runtime/worker.mjs"
+main = "../../build/worker.entrypoint.mjs"
 compatibility_date = "2026-04-15"
 compatibility_flags = ["nodejs_compat"]
