+++ examples/minimal-sinatra-with-db/Gemfile
@@ -4,6 +4,7 @@ source 'https://rubygems.org'
 
 ruby '>= 3.4.0'
 
+# Mode A — monorepo dev | Mode B — local `.gem` install (order: opal → runtime → sinatra → sequel-d1).
 gem 'opal', path: '../../vendor/opal-gem'
 gem 'cloudflare-workers-runtime', path: '../../gems/cloudflare-workers-runtime'
 gem 'sinatra-cloudflare-workers', path: '../../gems/sinatra-cloudflare-workers'
