+++ gems/cloudflare-workers-runtime/README.md
@@ -1,39 +1,36 @@
 # cloudflare-workers-runtime
 
-**Phase 15-B** — Core Ruby + Module Worker glue for [Opal](https://opalrb.com/) on [Cloudflare Workers](https://developers.cloudflare.com/workers/). This gem does **not** depend on Sinatra or Sequel; those live in separate gems (homurabi roadmap Phase 15-C / 15-D).
+Core Ruby + Module Worker glue for [Opal](https://opalrb.com/) on [Cloudflare Workers](https://developers.cloudflare.com/workers/). This gem does **not** depend on Sinatra or Sequel; those live in sibling gems.
 
 ## What you get
 
 - `require 'opal_patches'` — additive patches for Opal corelib vs real-world gems (Rack, etc.).
-- `require 'cloudflare_workers'` — `Rack::Handler::CloudflareWorkers`, `CloudflareWorkers.register_rack` / `register_scheduled` / `register_queue`, and `Cloudflare::D1Database`, `KVNamespace`, `R2Bucket`, `AI`, HTTP, multipart, cache, queue, durable object helpers.
-- `runtime/worker.mjs` — Module Worker that prefers `globalThis.__OPAL_WORKERS__` and falls back to legacy `__HOMURABI_*` globals (see `docs/TOOLCHAIN_CONTRACT.md` in the homurabi repo).
-- `runtime/setup-node-crypto.mjs` — exposes `node:crypto` on `globalThis` before the Opal bundle loads (matches Workers `nodejs_compat`).
+- `require 'cloudflare_workers'` — Rack handler, Cloudflare bindings, multipart, queue, Durable Objects, etc.
+- `runtime/worker_module.mjs` — fetch / scheduled / queue / DO adapters (**no Opal bundle import**).
+- `runtime/worker.mjs` — thin bootstrap (crypto shim → bundle → `worker_module`) for legacy layouts.
+- `runtime/setup-node-crypto.mjs` — `node:crypto` on `globalThis` before the Opal bundle loads.
+- `bin/cloudflare-workers-build` — single build pipeline (ERB → assets → Opal → patch → `worker.entrypoint.mjs`). Use `--standalone` in generated apps.
+- `docs/ARCHITECTURE.md` — wrangler `main`, codegen entrypoint, and fixed-import policy.
 
-## Minimal integration (Rack only, no Sinatra)
+## Quick start (homurabi monorepo)
 
-1. Add to your `Gemfile`:
+1. `Gemfile`: `gem 'cloudflare-workers-runtime', path: 'gems/cloudflare-workers-runtime'` and `gem 'opal', '= 1.8.3.rc1'` (path or exact pin).
+2. `bundle exec cloudflare-workers-build` — writes `build/hello.no-exit.mjs` and `build/worker.entrypoint.mjs`.
+3. `wrangler.toml`: `main = "build/worker.entrypoint.mjs"`, `compatibility_flags = ["nodejs_compat"]`.
 
-   ```ruby
-   gem 'opal', '~> 1.8.3.rc1'
-   gem 'cloudflare-workers-runtime', path: 'gems/cloudflare-workers-runtime' # or git / rubygems later
-   ```
+## Support matrix (indicative)
 
-2. Build an Opal ESM bundle that pre-requires patches and runtime:
+| Component   | Verified (dev)   |
+|------------|------------------|
+| Ruby       | 3.4.x            |
+| Node       | ≥ 20             |
+| Wrangler   | ^3.99            |
+| Opal       | = 1.8.3.rc1      |
 
-   ```text
-   bundle exec opal -c -E --esm --no-source-map \
-     -I lib -I vendor -I build \
-     -r opal_patches -r cloudflare_workers \
-     -o build/app.mjs config.ru # or your entry
-   ```
+## Wrangler config
 
-3. Point `wrangler.toml` at this gem’s worker and mirror the `import` path to your bundle (see homurabi’s `wrangler.toml` `main = "gems/cloudflare-workers-runtime/runtime/worker.mjs"` and the relative `../../../build/hello.no-exit.mjs` import inside `worker.mjs` for the monorepo layout).
-
-4. For **Sinatra** on Workers, use the upcoming `sinatra-cloudflare-workers` gem (Phase 15-C) instead of wiring everything by hand.
-
-## Wrangler template
-
-See `runtime/wrangler.toml.example` for optional D1 / KV / R2 / AI / DO / Queue snippets (opt-in comments).
+- **Supported as generated / documented**: `wrangler.toml` only.
+- `wrangler.json` / `wrangler.jsonc`: manual conversion only; not generated by this toolchain.
 
 ## License
