+++ gems/cloudflare-workers-runtime/exe/compile-assets
