+++ gems/cloudflare-workers-runtime/runtime/patch-opal-evals.mjs
