+++ gems/sinatra-cloudflare-workers/templates/project/app/app.rb.tt
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+require 'sinatra/cloudflare_workers'
+
+class App < Sinatra::Base
+  get '/' do
+    content_type 'text/plain; charset=utf-8'
+    '<%= app_name %>: ok'
+  end
+end
+
+run App
