+++ gems/sinatra-cloudflare-workers/templates/project/app/hello.rb.tt
@@ -1,0 +1,4 @@
+# await: call
+# frozen_string_literal: true
+# Opal entry — loaded with `-r` chain via cloudflare-workers-build.
+require 'app'
