+++ gems/sinatra-cloudflare-workers/templates/project/config.ru.tt
@@ -1,0 +1,2 @@
+# frozen_string_literal: true
+require_relative 'app/app'
