+++ gems/sinatra-cloudflare-workers/templates/project/views/index.erb
@@ -1,0 +1,1 @@
+<p>ok</p>
