+++ gems/sinatra-cloudflare-workers/templates/project/wrangler.toml.tt
@@ -1,0 +1,4 @@
+name = "<%= app_name %>"
+main = "worker.entrypoint.mjs"
+compatibility_date = "<%= compat_date %>"
+compatibility_flags = ["nodejs_compat"]
