+++ package.json
@@ -5,16 +5,16 @@
   "type": "module",
   "description": "Real Ruby + Sinatra on Cloudflare Workers via Opal",
   "scripts": {
-    "build:erb": "ruby bin/compile-erb --input views --output build/homurabi_templates.rb --namespace HomurabiTemplates",
-    "build:assets": "ruby bin/compile-assets --input public --output build/homurabi_assets.rb --namespace HomurabiAssets",
-    "build:opal": "ruby bin/inline-routes-for-opal && OPAL_PREFORK_DISABLE=1 bundle exec opal -c -E --esm --no-source-map -I app -I gems/cloudflare-workers-runtime/lib -I gems/sinatra-cloudflare-workers/lib -I gems/sequel-d1/lib -I lib -I vendor -I build -r opal_patches -r cloudflare_workers -r homurabi_templates -r homurabi_assets -o \"${HOMURABI_OPAL_OUTPUT:-build/hello.no-exit.mjs}\" \"${HOMURABI_OPAL_INPUT:-app/hello.rb}\" 2>build/opal.stderr.log && node bin/patch-opal-evals.mjs \"${HOMURABI_OPAL_PATCH_INPUT:-${HOMURABI_OPAL_OUTPUT:-build/hello.no-exit.mjs}}\"",
-    "build": "npm run build:erb && npm run build:assets && npm run build:opal",
+    "build:erb": "ruby gems/cloudflare-workers-runtime/exe/compile-erb --input views --output build/homurabi_templates.rb --namespace HomurabiTemplates",
+    "build:assets": "ruby gems/cloudflare-workers-runtime/exe/compile-assets --input public --output build/homurabi_assets.rb --namespace HomurabiAssets",
+    "build:opal": "bundle exec cloudflare-workers-build",
+    "build": "bundle exec cloudflare-workers-build",
     "dev": "npm run build && wrangler dev --port 8787 --ip 127.0.0.1",
     "dev:dlq": "npm run build && wrangler dev --env dlq-demo --port 8787 --ip 127.0.0.1",
     "d1:init": "bin/init-local-d1",
     "test:workers": "bin/test-on-workers",
     "deploy": "npm run build && wrangler deploy",
-    "build:test": "ruby bin/compile-erb && OPAL_PREFORK_DISABLE=1 bundle exec opal -c -E --esm --no-source-map -I app -I gems/cloudflare-workers-runtime/lib -I gems/sinatra-cloudflare-workers/lib -I gems/sequel-d1/lib -I lib -I vendor -I build -r opal_patches -r cloudflare_workers -r homurabi_templates -o build/smoke-test.mjs test/smoke.rb 2>build/opal.stderr.log",
+    "build:test": "ruby gems/cloudflare-workers-runtime/exe/compile-erb --input views --output build/homurabi_templates.rb --namespace HomurabiTemplates && OPAL_PREFORK_DISABLE=1 bundle exec opal -c -E --esm --no-source-map -I app -I gems/cloudflare-workers-runtime/lib -I gems/sinatra-cloudflare-workers/lib -I gems/sequel-d1/lib -I lib -I vendor -I build -r opal_patches -r cloudflare_workers -r homurabi_templates -o build/smoke-test.mjs test/smoke.rb 2>build/opal.stderr.log && node gems/cloudflare-workers-runtime/runtime/patch-opal-evals.mjs build/smoke-test.mjs",
     "test": "npm run build:test && node --import ./gems/cloudflare-workers-runtime/runtime/setup-node-crypto.mjs build/smoke-test.mjs && npm run test:http && npm run test:crypto && npm run test:jwt && npm run test:scheduled && npm run test:ai && npm run test:faraday && npm run test:multipart && npm run test:streaming && npm run test:octokit && npm run test:do && npm run test:cache && npm run test:queue && npm run test:sequel && npm run test:fiber-await && npm run test:classic-sinatra",
     "build:classic-sinatra-test": "OPAL_PREFORK_DISABLE=1 bundle exec opal -c -E --esm --no-source-map -I app -I gems/cloudflare-workers-runtime/lib -I gems/sinatra-cloudflare-workers/lib -I gems/sequel-d1/lib -I lib -I vendor -I build -r opal_patches -o build/classic-sinatra-smoke.mjs test/classic_sinatra_smoke.rb 2>build/opal.stderr.log",
     "test:classic-sinatra": "npm run build:classic-sinatra-test && node --import ./gems/cloudflare-workers-runtime/runtime/setup-node-crypto.mjs build/classic-sinatra-smoke.mjs",
